requirements.txt
opencv-python
numpy
flask
pillow
rembg
psutil
keyboard